+++ src/pytest_bluezenv/__init__.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8; mode: python; eval: (blacken-mode); -*-
 # SPDX-License-Identifier: GPL-2.0-or-later
 
-from .rpc import RemoteError
+from .rpc import RemoteError, RemoteTimeoutError
 
 from .env import *
 from .utils import *
 from .host_plugins import *
 from .btmon import *
 from .agent import *
+from .le import *
 
 from .plugin import *